tutorial: add QEMU details
Recently, I had to dig into how to set up QEMU with a virtualized TPM2.0
device to debug a potential issue in the in-kernel resource manager
(/dev/tpmrm0). Their is a quite a bit of information on setting up QEMU,
and a TPM device, it seemed to be lacking in setting up the TPM2.0
device. The crux was the subtle and easy to overlook --tpm2 option
to the SWTPM.

Build a simple tutorial to highlight a setup using TPM2.0.

Signed-off-by: William Roberts <william.c.roberts@intel.com>